--- src/main/java/com/example/authenticationserver/test/JwksRepository.java
@@ -1,8 +1,0 @@
-package com.example.authenticationserver.test;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface JwksRepository extends JpaRepository<JwksEntity,String> {
-}
